feat(ci): add workflows to provision runners
Rationale is to be able to provision self-hosted runners when a test run
is needed.

Due to nested virtualization, virtualbox and Vagrant, we need a very
specific setup to make these run.

Purpose is to create the self-hosted runners prior to the run, with a
few extra workflows to clean-up potential runners that were left behind
in prior runs.